// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/DocumentDetails.ts
export interface DocumentDetails {
    /**
     * 
     * @type {string}
     * @memberof DocumentDetails
     */
    documentId?: string;
    /**
     * 
     * @type {string}
     * @memberof DocumentDetails
     */
    signingStatus?: DocumentDetailsSigningStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof DocumentDetails
     */
    signedDocumentType?: DocumentDetailsSignedDocumentTypeEnum;
}


/**
 * @export
 */
export const DocumentDetailsSigningStatusEnum = {
    Pending: 'PENDING',
    Completed: 'COMPLETED',
    Failed: 'FAILED'
} as const;
export type DocumentDetailsSigningStatusEnum = typeof DocumentDetailsSigningStatusEnum[keyof typeof DocumentDetailsSigningStatusEnum];

/**
 * @export
 */
export const DocumentDetailsSignedDocumentTypeEnum = {
    Fixed: 'FIXED',
    FixedCustomizable: 'FIXED_CUSTOMIZABLE',
    LocDraw: 'LOC_DRAW',
    LineOfCredit: 'LINE_OF_CREDIT',
    FixedTermContract: 'FIXED_TERM_CONTRACT',
    FixedTermContractUsd: 'FIXED_TERM_CONTRACT_USD',
    FixedTermContractSecuredUsd: 'FIXED_TERM_CONTRACT_SECURED_USD',
    FixedTermContractAprEarlySecuredUsd: 'FIXED_TERM_CONTRACT_APR_EARLY_SECURED_USD',
    FixedTermContractAprEarlyUsd: 'FIXED_TERM_CONTRACT_APR_EARLY_USD',
    FixedTermContractUkPlus: 'FIXED_TERM_CONTRACT_UK_PLUS',
    FixedTermContractDeu: 'FIXED_TERM_CONTRACT_DEU',
    FixedTermContractCanadaUnsecured: 'FIXED_TERM_CONTRACT_CANADA_UNSECURED',
    FixedTermContractCanadaSecured: 'FIXED_TERM_CONTRACT_CANADA_SECURED',
    FixedTermContractBespokeUnsecured: 'FIXED_TERM_CONTRACT_BESPOKE_UNSECURED',
    FixedTermContractBespokeSecured: 'FIXED_TERM_CONTRACT_BESPOKE_SECURED',
    LocFrameworkUsUnsecured: 'LOC_FRAMEWORK_US_UNSECURED',
    LocFrameworkUsSecured: 'LOC_FRAMEWORK_US_SECURED',
    LocFrameworkUk: 'LOC_FRAMEWORK_UK',
    Loc20FrameworkUsSecured: 'LOC_20_FRAMEWORK_US_SECURED',
    Loc20CaliforniaDisclosure: 'LOC_20_CALIFORNIA_DISCLOSURE',
    Loc20RefiCaliforniaDisclosure: 'LOC_20_REFI_CALIFORNIA_DISCLOSURE',
    DisclosureCalifornia: 'DISCLOSURE_CALIFORNIA',
    DisclosureNewYork: 'DISCLOSURE_NEW_YORK',
    DisclosureGeorgiaFlorida: 'DISCLOSURE_GEORGIA_FLORIDA',
    DisclosureKansas: 'DISCLOSURE_KANSAS',
    DisclosureConnecticut: 'DISCLOSURE_CONNECTICUT',
    DisclosureUtah: 'DISCLOSURE_UTAH',
    DisclosureVirginia: 'DISCLOSURE_VIRGINIA',
    UniformCommercialCode: 'UNIFORM_COMMERCIAL_CODE',
    PersonalGuarantee: 'PERSONAL_GUARANTEE',
    Debenture: 'DEBENTURE',
    DeedsOfPriority: 'DEEDS_OF_PRIORITY',
    CrossCorporateSingleGuarantee: 'CROSS_CORPORATE_SINGLE_GUARANTEE',
    CrossCorporateMultipleGuarantee: 'CROSS_CORPORATE_MULTIPLE_GUARANTEE',
    PersonalPropertySecurityAct: 'PERSONAL_PROPERTY_SECURITY_ACT',
    RbfContractUsa: 'RBF_CONTRACT_USA',
    RbfContractUk: 'RBF_CONTRACT_UK',
    RbfContractCanada: 'RBF_CONTRACT_CANADA',
    RbfDisclosureCalifornia: 'RBF_DISCLOSURE_CALIFORNIA',
    RbfDisclosureNewYork: 'RBF_DISCLOSURE_NEW_YORK',
    RbfDisclosureFlorida: 'RBF_DISCLOSURE_FLORIDA',
    RbfDisclosureGeorgia: 'RBF_DISCLOSURE_GEORGIA',
    RbfDisclosureKansas: 'RBF_DISCLOSURE_KANSAS',
    RbfDisclosureLouisiana: 'RBF_DISCLOSURE_LOUISIANA',
    RbfDisclosureMissouri: 'RBF_DISCLOSURE_MISSOURI',
    RbfDisclosureUtah: 'RBF_DISCLOSURE_UTAH',
    RbfDisclosureVirginia: 'RBF_DISCLOSURE_VIRGINIA',
    RbfDisclosureConnecticut: 'RBF_DISCLOSURE_CONNECTICUT',
    DailyPayoutContract: 'DAILY_PAYOUT_CONTRACT',
    Unknown: 'UNKNOWN'
} as const;
export type DocumentDetailsSignedDocumentTypeEnum = typeof DocumentDetailsSignedDocumentTypeEnum[keyof typeof DocumentDetailsSignedDocumentTypeEnum];